chore: move gnodev to `/contribs` folder
