+++ contribs/gnodev/Makefile
@@ -1,0 +1,9 @@
+GNOROOT_DIR ?= $(abspath $(lastword $(MAKEFILE_LIST))/../../../)
+
+GOBUILD_FLAGS := -ldflags "-X github.com/gnolang/gno/gnovm/pkg/gnoenv._GNOROOT=$(GNOROOT_DIR)"
+
+install:
+	go install $(GOBUILD_FLAGS) ./cmd/gnodev
+
+build:
+	go build $(GOBUILD_FLAGS) -o build/gnodev ./cmd/gno
